// src/types/nameEventSocket.ts
export enum ENameEvent {
  JOIN_GROUP = 'joinGroup',
  NEW_GROUP = 'newGroup',
  USER_ONLINE = 'userOnline',
  UPDATE_ONLINE_USER = 'updateOnlineUsers',
  NEW_NOTIFICATION = 'newNotification',
  EDIT_GROUP = 'editGroup',
  EDIT_MESSAGE = 'editMessage',
  RECEIVE_MESSAGE = 'receiveMessage',
  KICK_USER = 'kickUserFromGroup',
  ADD_MEMBER = 'addMemberFromGroup',
  CHANGE_THEME = 'changeTheme',
  DELETE_MESSAGE = 'deleteMessage',
  DELETE_MESSAGE_ME = 'deleteMessageForMe',
  REACT_MESSAGE = 'reactMessage',
  LEAVE_GROUP = 'leaveGroup',
  DISCONNECT = 'disconnect'
}